test: split shared script spine helpers
